backend/gpubroker/requirements.txt
# Django core
Django>=5.0,<6.0
django-ninja>=1.0,<2.0
channels>=4.0,<5.0
channels-redis>=4.0,<5.0
daphne>=4.0,<5.0

# Database
psycopg[binary]>=3.0,<4.0
dj-database-url>=2.0,<3.0

# Authentication
python-jose[cryptography]>=3.3,<4.0
passlib[argon2]>=1.7,<2.0

# Caching
django-redis>=5.0,<6.0

# Observability
django-prometheus>=2.0,<3.0
opentelemetry-instrumentation-django>=0.40b0
python-json-logger>=2.0,<3.0

# Security
django-cors-headers>=4.0,<5.0
django-ratelimit>=4.0,<5.0

# Payments
stripe>=7.0,<8.0

# Utilities
django-environ>=0.11,<1.0
whitenoise>=6.0,<7.0
httpx>=0.25,<1.0
requests>=2.31,<3.0
email-validator>=2.0,<3.0
numpy>=1.24,<2.0
pandas>=2.0,<3.0

